Move unit test of util to correct file
